Refactored Test cases for GenericIPAddress Field
